Couldn't see anything that was pre-coded to deploy to testnet so made some tweaks/added a deploy_testnet.js file for easy testnet deploy
